badgesfab/requirements.txt
Django==1.9.7
django-activity-stream==0.6.2
django-bootstrap3==7.0.1
django-crispy-forms==1.6.0
django-filter==0.13.0
django-jsonfield==1.0.0
django-money==0.8
django-taggit==0.20.1
djangorestframework==3.3.3
factory-boy==2.7.0
fake-factory==0.5.7
mysqlclient==1.3.7
Pillow==2.4.0
py-moneyed==0.6.0
python-dateutil==2.5.3
pytz==2016.4
six==1.10.0
django-predicate

#TODO: purge requirements before sending